+++ src/constants/signals.js
@@ -1,0 +1,3 @@
+export const SIGNAL_TYPES = {
+  SERVICE_LEVEL: 'service_level',
+};