+++ src/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useFetchFlows } from './use-fetch-flows';
 export { default as useServiceLevelsSearch } from './use-service-levels-search';
-export { default as nrqlQueries } from './fetch-kpi-values';
+export { default as useFetchServiceLevels } from './use-fetch-service-levels';
+export { default as useFetchKpis } from './use-fetch-kpis';